fix: unsubscribe from notifications when server is deleted, move resetServer
